fix difference in sign extension in unix vs plan 9
This fixes the audio

+++ src/Sound.cpp
@@ -335,9 +335,9 @@ int MakePixToneObject(const PIXTONEPARAMETER *ptp, int ptp_num, int no)
 
 		for (j = 0; j < ptp_pointer->size; j++)
 		{
-			if (pcm_buffer[j] + mixed_pcm_buffer[j] - 0x100 < -0x7F)
+			if ((int)(pcm_buffer[j] + mixed_pcm_buffer[j]) - 0x100 < -0x7F)
 				mixed_pcm_buffer[j] = 0;
-			else if (pcm_buffer[j] + mixed_pcm_buffer[j] - 0x100 > 0x7F)
+			else if ((int)(pcm_buffer[j] + mixed_pcm_buffer[j]) - 0x100 > 0x7F)
 				mixed_pcm_buffer[j] = 0xFF;
 			else
 				mixed_pcm_buffer[j] = mixed_pcm_buffer[j] + pcm_buffer[j] - 0x80;
